--- frontend/packages/schema-editor/src/components/SchemaInspector/ItemFieldsTab.module.css
@@ -1,13 +1,0 @@
-.root {
-  display: grid;
-  gap: 1rem;
-  grid-template-columns: 3fr 2fr auto auto;
-}
-
-.addButtonCell {
-  grid-column: 1 / 5;
-}
-
-.addButtonCell > * {
-  width: 100%;
-}
