+++ frontend/packages/schema-editor/src/components/SchemaInspector/ItemFieldsTab/ItemFieldsTab.module.css
@@ -1,0 +1,5 @@
+.root {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
