+++ frontend/packages/schema-editor/src/components/SchemaInspector/ItemFieldsTab/ItemFieldsTable/ItemFieldsTable.module.css
@@ -1,0 +1,9 @@
+.tableHeaderLeft {
+  text-align: left;
+  padding-bottom: 1rem;
+}
+
+.tableHeaderCenter {
+  text-align: center;
+  padding-bottom: 1rem;
+}
