+++ frontend/packages/schema-editor/src/components/SchemaInspector/ItemFieldsTab/ItemFieldsTable/ItemFieldsTableRow/ItemFieldsTableRow.module.css
@@ -1,0 +1,30 @@
+.tableData {
+  vertical-align: middle;
+}
+
+.tableDataContentWrapper {
+  height: 50px;
+}
+
+.tableDataContentWrapperSelect {
+  height: 50px;
+}
+
+.tableDataContentWrapperSwitch {
+  height: 50px;
+}
+
+.tableData1 {
+  width: 40%;
+  text-align: left;
+}
+.tableData2 {
+  width: 30%;
+  text-align: left;
+}
+.tableData3 {
+  width: 10%;
+}
+.tableData4 {
+  width: 10%;
+}
