+++ frontend/packages/schema-editor/src/components/SchemaInspector/ItemFieldsTab/ItemFieldsTable/ItemFieldsTableRow/index.ts
@@ -1,0 +1,1 @@
+export { ItemFieldsTableRow } from './ItemFieldsTableRow';
