+++ frontend/packages/schema-editor/src/components/SchemaInspector/ItemFieldsTab/ItemFieldsTable/index.ts
@@ -1,0 +1,1 @@
+export { ItemFieldsTable } from './ItemFieldsTable';
