+++ frontend/packages/schema-editor/src/components/SchemaInspector/ItemFieldsTab/index.ts
@@ -1,0 +1,1 @@
+export { ItemFieldsTab } from './ItemFieldsTab';
