--- frontend/packages/schema-editor/src/components/SchemaInspector/PropertyItem.module.css
@@ -1,20 +1,0 @@
-.gridItem {
-  place-self: start center;
-}
-
-.nameInputCell {
-  place-self: start stretch;
-}
-
-.nameInputCell > * {
-  width: 100%;
-  height: 100%;
-}
-
-.typeSelectCell {
-  justify-self: stretch;
-}
-
-.requiredCheckCell {
-  padding: 6px;
-}
